+++ .gitignore
@@ -1,0 +1,3 @@
+config.php
+logs/errors/major.txt
+logs/errors/minor.txt
