--- logs/errors/major.txt
@@ -1,3 +1,0 @@
-["E_RECOVERABLE_ERROR","Uncaught PDOException: SQLSTATE[HY000] [1045] Access denied for user ''@'localhost' (using password: NO) in \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/class\/db.php:11\nStack trace:\n#0 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/class\/db.php(11): PDO->__construct('mysql:host=loca...', '', '', Array)\n#1 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/dbconn.php(5): glPDO->__construct('mysql:host=loca...', '', '')\n#2 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/init.php(21): include('\/Users\/cday\/Pro...')\n#3 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/index.php(5): require('\/Users\/cday\/Pro...')\n#4 {main}\n  thrown","\/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/class\/db.php",11,"2023-05-30 16:03:55"]
-["E_RECOVERABLE_ERROR","Uncaught PDOException: SQLSTATE[HY000] [1045] Access denied for user ''@'localhost' (using password: NO) in \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/class\/db.php:11\nStack trace:\n#0 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/class\/db.php(11): PDO->__construct('mysql:host=loca...', '', '', Array)\n#1 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/dbconn.php(5): glPDO->__construct('mysql:host=loca...', '', '')\n#2 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/init.php(21): include('\/Users\/cday\/Pro...')\n#3 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/index.php(5): require('\/Users\/cday\/Pro...')\n#4 {main}\n  thrown","\/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/class\/db.php",11,"2023-05-30 16:12:53"]
-["E_RECOVERABLE_ERROR","Uncaught PDOException: SQLSTATE[HY000] [1045] Access denied for user ''@'localhost' (using password: NO) in \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/class\/db.php:11\nStack trace:\n#0 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/class\/db.php(11): PDO->__construct('mysql:host=loca...', '', '', Array)\n#1 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/dbconn.php(5): glPDO->__construct('mysql:host=loca...', '', '')\n#2 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/init.php(21): include('\/Users\/cday\/Pro...')\n#3 \/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/index.php(5): require('\/Users\/cday\/Pro...')\n#4 {main}\n  thrown","\/Users\/cday\/Projects\/GLScript\/release\/openPBBG\/class\/db.php",11,"2023-05-30 16:12:54"]
